Checkpoint compilation work
Nothing is compiling yet but the basic approach is in place.

+++ fea-rs/src/types/metrics.rs
@@ -26,8 +26,8 @@ pub enum ValueRecord {
 
 #[derive(Clone, Debug)]
 pub enum Anchor {
-    Coord { x: i32, y: i32 },
-    Contour { x: i32, y: i32, point: u32 },
+    Coord { x: i16, y: i16 },
+    Contour { x: i16, y: i16, point: u16 },
     //un
     //Device { x: i32, y: i32, x_dev: (), y_dev: (), },
     //unimplemented
